requirements.txt
pygame
pillow
